requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.4.0
autopep8==1.6.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==5.0.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.0
channels==4.1.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
cssselect2==0.7.0
cupshelpers==1.0
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.4
django-check-seo==0.6.2
django-ckeditor==6.7.1
django-classy-tags==4.0.0
django-cms==3.11.3
django-cors-headers==4.2.0
django-filer==2.2.5
django-formtools==2.4.1
django-jazzmin==3.0.1
django-js-asset==2.1.0
django-meta==2.2.0
django-mptt==0.14.0
django-plotly-dash==2.4.3
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-treebeard==4.7
djangocms-admin-style==3.2.4
djangocms-page-meta==1.2.0
djangorestframework==3.14.0
dpd-components==0.1.0
duplicity==0.8.21
easy-thumbnails==2.8.5
entrypoints==0.4
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.7.0
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.47.0
h5py==3.7.0
httplib2==0.22.0
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
monotonic==1.6
more-itertools==8.10.0
multidict==5.1.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==2.1.2
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==23.1
pandas==2.2.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2022.12.19
platformdirs==2.5.2
plotly==5.24.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.24.1
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.42.3
proton-vpn-gtk-app==4.9.5
proton-vpn-network-manager==0.12.13
psutil==5.9.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.18.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.7.1
roman-nepali-translator==0.1.1
rsa==4.9
scipy==1.14.1
SecretStorage==3.3.1
Send2Trash==1.8.0
sentry-sdk==1.4.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.2.0
statsmodels==0.14.4
svglib==1.5.1
systemd-python==234
tenacity==9.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.1
tqdm==4.65.0
traitlets==5.2.1.post0
typing_extensions==4.7.1
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
Unidecode==1.1.2
uritemplate==4.1.1
urllib3==1.26.16
usb-creator==0.3.7
vimeo-downloader==0.4.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
whitenoise==6.6.0
widgetsnbextension==3.6.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
